scripts/core/gpg_manager/requirements.txt
Flask==2.3.3
python-gnupg==0.5.1
python-dotenv==1.0.0
gunicorn==21.2.0
